apps/api/airm/pyproject.toml
# Copyright © Advanced Micro Devices, Inc., or its affiliates.
#
# SPDX-License-Identifier: MIT

[project]
name = "api"
version = "0.1.0"
description = "AMD Resource Manager API"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "asyncpg~=0.30.0",
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.4.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "prometheus-api-client>=0.5.7",
    "PyYAML>=6.0.2",
    "python-dotenv>=1.1.0",
    "python-keycloak>=5.5.0",
    "sqlalchemy[asyncio]~=2.0.36",
    "uvicorn[standard]>=0.34.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "aio-pika>=9.5.5",
]

[dependency-groups]
dev = [
    "async-asgi-testclient>=1.4.11",
    "coverage>=7.8.0",
    "debugpy>=1.8.0",
    "docker>=7.1.0",
    "filelock>=3.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.1.1",
    "testcontainers[postgres,rabbitmq]>=4.14.0",
    "pytest-env>=1.1.5",
    "pytest-httpx>=0.35.0",
    "pytest-xdist>=3.6.0",
    "pytest-loguru>=0.4.0",
    "responses>=0.25.7",
    "types-docker>=7.1.0.20250705",
    "robotframework>=7.1.1",
    "robotframework-requests>=0.9.7",
    "requests>=2.32.3",
    "certifi>=2025.1.31",
    "PyJWT>=2.10.1",
    "tenacity>=9.1.2",
]
